feat: rename +UrlDisplayer to FeedTitleDisplayer
