+++ frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/components/FeedItems/components/FeedTitleDisplayer/FeedTitleDisplayer.tsx
@@ -1,0 +1,37 @@
+import { useState } from "react";
+
+import { enhanceUrlDisplay } from "./utils/enhanceUrlDisplay";
+
+type FeedTitleDisplayerProps = {
+  title: string;
+  truncate?: boolean;
+};
+
+export const FeedTitleDisplayer = ({
+  title,
+  truncate = false,
+}: FeedTitleDisplayerProps): JSX.Element => {
+  const [showFullUrl, setShowFullUrl] = useState(false);
+
+  const toggleShowFullUrl = () => {
+    setShowFullUrl(!showFullUrl);
+  };
+
+  if (truncate) {
+    return (
+      <div className="overflow-hidden">
+        <span className="cursor-pointer" onClick={toggleShowFullUrl}>
+          <p className={showFullUrl ? "" : "truncate"}>{title}</p>
+        </span>
+      </div>
+    );
+  }
+
+  return (
+    <div>
+      <span className="cursor-pointer" onClick={toggleShowFullUrl}>
+        {showFullUrl ? title : enhanceUrlDisplay(title)}
+      </span>
+    </div>
+  );
+};
