--- frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/components/FeedItems/components/UrlDisplay/index.ts
@@ -1,1 +1,0 @@
-export * from "./UrlDisplay";
